refactor(core): moved core to server package
BREAKING CHANGE: Core will no longer be built into a jar so Babblebot binary files will need to
remove it from the classpath

#78
