--- core/build.gradle
@@ -1,18 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(":api")
-    implementation 'org.projectlombok:lombok:1.18.16'
-    implementation group: 'org.slf4j', name: 'slf4j-api', version: '1.7.30'
-    annotationProcessor 'org.projectlombok:lombok:1.18.12'
-    implementation 'com.google.code.gson:gson:2.8.6'
-    compile 'com.github.zafarkhaja:java-semver:0.9.0'
-    implementation 'commons-io:commons-io:2.8.0'
-
-}
