+++ gradle.properties
@@ -1,2 +1,2 @@
-version = 2.1.5
+version = 2.1.6
 systemProp.org.gradle.internal.publish.checksums.insecure = true
