+++ server/src/main/java/uk/co/bjdavies/core/CorePlugin.java
@@ -30,6 +30,8 @@ import discord4j.core.event.domain.lifecycle.ReadyEvent;
 import discord4j.core.object.entity.channel.TextChannel;
 import discord4j.core.spec.EmbedCreateSpec;
 import lombok.extern.slf4j.Slf4j;
+import net.bdavies.db.Model;
+import net.bdavies.db.WhereStatement;
 import reactor.core.publisher.Mono;
 import uk.co.bjdavies.api.IApplication;
 import uk.co.bjdavies.api.command.Command;
@@ -37,8 +39,6 @@ import uk.co.bjdavies.api.command.CommandParam;
 import uk.co.bjdavies.api.command.ICommandContext;
 import uk.co.bjdavies.api.command.ICommandDispatcher;
 import uk.co.bjdavies.api.config.IDiscordConfig;
-import uk.co.bjdavies.api.db.Model;
-import uk.co.bjdavies.api.db.WhereStatement;
 import uk.co.bjdavies.api.discord.IDiscordFacade;
 import uk.co.bjdavies.api.plugins.IPluginEvents;
 import uk.co.bjdavies.api.plugins.IPluginSettings;
